Extract CLI output rendering into scripts/cli_rendering.py
scripts/cli.py mixed argument parsing/dispatch with ~350 lines of pure
output formatting (table/csv/curator_desk_csv/xml renderers and their
helpers), all of which depend only on stdlib imports and each other -
nothing from BioAnalyzerCLI or argument parsing. Moved verbatim into a
new module; cli.py now imports render_results, render_retrieval, and
ANALYSIS_FIELDS (the only names referenced outside the moved block).
No behavior change - confirmed via the existing curator-desk CSV test,
a CLI --help smoke test, and black --check.
